Fix Homebrew zsh command completion issue
For whatever reason, `brew update` breaks my zsh's command
completion setup.

I feel that this is a bug with Homebrew, but I also feel that
it will take a lot of time figuring out what is actually causing the
problem, and even more to fix it.
So, instead, we ensure that the symbolic links that Homebrew creates
point to the correct files (instead of relative paths that Homebrew
likes).